requirements.txt
fastapi
uvicorn
redis
autogen-agentchat
autogen-ext[openai]
python-dotenv
pydantic